separate structs into seperate files and fix bug in executionview where video doesn't load when next
