--- WorkItOut/Core/Utilities/Data/Entities/YogaHistory.swift
@@ -1,9 +1,0 @@
-//
-//  YogaHistory.swift
-//  Mamaste
-//
-//  Created by Elvin Sestomi on 21/11/23.
-//
-
-import Foundation
-
